Rework sending of TCP data
When the socket buffer becomes full, the old behaviour was to drop
subsequent segments, causing TCP retransmissions which heavily affected
the upload bandwidth. Now all the segments are buffered in memory until
the socket becomes available. Moreover, sends are now processed in
batches, reducing the number ACKs (less writes to the VPN fd).
The only drawback is an higher memory usage.